+++ libs/core/src/handlers/events.ts
@@ -6,7 +6,7 @@ export enum SeelenEvent {
   HandleLayeredHitboxes = "handle-layered",
   SystemMonitorsChanged = "system::monitors-changed",
   SystemLanguagesChanged = "system::languages-changed",
-  SystemBrightnessChanged = "system::brightness-changed",
+  SystemMonitorsBrightnessChanged = "system::monitors-brightness-changed",
   UserChanged = "user-changed",
   UserFolderChanged = "user::known-folder-changed",
   UserAppWindowsChanged = "user::windows-changed",