+++ src/background/exposed.rs
@@ -263,6 +263,7 @@ pub fn register_invoke_handler(app_builder: Builder<Wry>) -> Builder<Wry> {
     use crate::modules::apps::infrastructure::*;
     use crate::modules::media::devices::infrastructure::*;
     use crate::modules::media::players::infrastructure::*;
+    use crate::modules::monitors::brightness::infrastructure::*;
     use crate::modules::monitors::infrastructure::*;
     use crate::modules::network::infrastructure::*;
     use crate::modules::notifications::infrastructure::*;