+++ src/background/modules/monitors/brightness/mod.rs
@@ -1,6 +1,3 @@
 mod application;
-pub mod brightness;
 mod domain;
 pub mod infrastructure;
-
-pub use application::*;