+++ src/background/modules/monitors/infrastructure.rs
@@ -1,17 +1,11 @@
 use std::sync::Once;
 
-use seelen_core::{
-    handlers::SeelenEvent,
-    system_state::{Brightness, PhysicalMonitor},
-};
+use seelen_core::{handlers::SeelenEvent, system_state::PhysicalMonitor};
 
 use crate::{
-    app::emit_to_webviews,
-    error::Result,
-    windows_api::{monitor::Monitor, MonitorEnumerator},
+    app::emit_to_webviews, error::Result, modules::monitors::MonitorManager,
+    windows_api::MonitorEnumerator,
 };
-
-use super::MonitorManager;
 
 fn get_monitor_manager() -> &'static MonitorManager {
     static TAURI_EVENT_REGISTRATION: Once = Once::new();
@@ -34,28 +28,4 @@ pub fn get_connected_monitors() -> Result<Vec<PhysicalMonitor>> {
         monitors.push(m.try_into()?);
     }
     Ok(monitors)
-}
-
-#[tauri::command(async)]
-pub fn get_main_monitor_brightness() -> Result<Option<Brightness>> {
-    let monitor = Monitor::primary();
-    let device = monitor.as_monitor_view()?.primary_target()?;
-
-    let current = device.ioctl_query_display_brightness()?;
-    if current == 0 && device.ioctl_set_display_brightness(0).is_err() {
-        return Ok(None);
-    }
-
-    Ok(Some(Brightness {
-        min: 0,
-        max: 100,
-        current: current as u32,
-    }))
-}
-
-#[tauri::command(async)]
-pub fn set_main_monitor_brightness(brightness: u8) -> Result<()> {
-    let monitor = Monitor::primary();
-    let device = monitor.as_monitor_view()?.primary_target()?;
-    device.ioctl_set_display_brightness(brightness.min(100))
 }