+++ changelog.md
@@ -45,6 +45,7 @@
 - run as admin not working on dock/taskbar.
 - single key shortcuts not taking in care holded keys.
 - wallpaper sometimes not showing on windows 11.
+- no emitting brightness events.
 
 ## [2.4.11]